gateway/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0
langsmith>=0.1.0
sse-starlette>=1.8.0
watchdog>=4.0.0
httpx>=0.27.0
apscheduler>=3.10.0
jinja2>=3.1.0
